add separate Descriptor class
