+++ zntrack/descriptor/__init__.py
@@ -1,0 +1,3 @@
+from .base import Descriptor
+
+__all__ = ["Descriptor"]
